Refactor MyArticles section to use @components
