+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/Header/Header.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 // components
-import Actions from './Actions';
-import Links from './Links';
+import Actions from './Actions/Actions.jsx';
+import Links from './Links/Links.jsx';
 
 const isEnglishWikipedia = ({ assignment, course }) => () => {
   const { language, project } = assignment;
