+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/Header/Links/Links.jsx
@@ -8,10 +8,10 @@ import PeerReviewLink from './PeerReviewLink';
 import EditorLink from './EditorLink';
 import ReviewerLink from './ReviewerLink';
 
-import Separator from '../../../../../../common/Separator';
+import Separator from '@components/overview/my_articles/common/Separator.jsx';
 
 // constants
-import { REVIEWING_ROLE } from '../../../../../../../../../constants/assignments';
+import { REVIEWING_ROLE } from '~/app/assets/javascripts/constants/assignments';
 
 // helper functions
 const interleaveSeparators = (acc, link, index, collection) => {
