+++ test/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/index.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import '../../../../../../../../testHelper';
 
-import ProgressTracker from '../../../../../../../../../app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker';
+import ProgressTracker from '@components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/ProgressTracker';
 
 describe('ProgressTracker', () => {
   const props = {
